Fix all warnings and errors in console [ci skip]
