+++ src/components/Layout/Sidebar/components/ActiveIndicator.style.ts
@@ -2,10 +2,10 @@ import { FONT_SIZE_LG, TRANSITION_NORMAL } from "@/styles/variables";
 import { convertHexToRGBA } from "@/utils/colors";
 import styled from "styled-components";
 
-export const StyledActiveIndicator = styled.div<{ linkTop: number }>`
+export const StyledActiveIndicator = styled.div<{ $pos: number }>`
     position: absolute;
     left: -14%;
-    top: ${({ linkTop }) => linkTop + 4}px;
+    top: ${({ $pos }) => $pos + 4}px;
 
     height: ${FONT_SIZE_LG};
     width: 13%;
