+++ src/components/Layout/Sidebar/components/NavLink.component.tsx
@@ -17,7 +17,6 @@ const NavLink = React.forwardRef<HTMLElement, NavLinkProps>(
 				to={to}
 				onClick={onClick}
 				tabIndex={tabIndex}
-				active={active}
 			>
 				{children}
 			</StyledNavLink>
