+++ src/types/store/theme.d.ts
@@ -55,7 +55,6 @@ type ThemeHeader = {
 type ThemeThemePage = {
 	iconColor: string;
 	iconHoverColor: string;
-	activeThemeColor: string;
 	selectedThemeColor: string;
 };
